background-agents/requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0.1
watchdog==3.0.0
schedule==1.2.0

# AI and ML
openai==1.3.0
anthropic==0.7.0
langchain==0.0.350
langchain-openai==0.0.2

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0
psutil==5.9.6

# Security and validation
bandit==1.7.5
safety==2.3.5
cryptography==41.0.7

# Code analysis
ast==0.0.2
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Performance monitoring
memory-profiler==0.61.0
py-spy==0.3.14

# Web framework for API
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0

# Database and storage
sqlite3
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development tools
pre-commit==3.5.0
black==23.11.0
isort==5.12.0
